Move bookInfo tests to separate file.
