+++ src/lib/__tests__/bookDataTransformerService.test.ts
@@ -1,37 +1,6 @@
-import { BookInfo, IBookInfo } from '../../models/bookInfo'
+import { BookInfo } from '../../models/bookInfo'
 import { BookDataTransformer,  }from '../bookDataTransformerService'
-import { sampleBookData } from '../fixtures/sampleBookData'
-
-const details: IBookInfo = {
-  authors: sampleBookData.items[0].volumeInfo.authors,
-  title: sampleBookData.items[0].volumeInfo.title,
-  publisher: sampleBookData.items[0].volumeInfo.publisher,
-  image: sampleBookData.items[0].volumeInfo.imageLinks.thumbnail,
-  url: sampleBookData.items[0].volumeInfo.infoLink
-}
-
-describe('BookInfo class structure', () => {
-  test("contains an author property", () => {
-    const info = new BookInfo(details)
-    expect(info).toHaveProperty('authors')
-  })
-  test("contains a title property", () => {
-    const info = new BookInfo(details)
-    expect(info).toHaveProperty('title')
-  })
-  test("contains a publishingCompany property", () => {
-    const info = new BookInfo(details)
-    expect(info).toHaveProperty('publisher')
-  })
-  test("contains an image property", () => {
-    const info = new BookInfo(details)
-    expect(info).toHaveProperty('image')
-  })
-  test("contains a url property", () => {
-    const info = new BookInfo(details)
-    expect(info).toHaveProperty('url')
-  })
-})
+import { sampleBookData } from '../../shared/fixtures/sampleBookData'
 
 describe('BookTransformer with successful search', () => {
   test('returns an array of BookInfo objects of the correct shape', () => {
@@ -80,12 +49,4 @@ describe('BookTransformer with successful search', () => {
     ]
     expect(bookData).toEqual(testData)
   })
-})
-
-describe('BookTransformer with unsuccessful search', () => {
-  // test('returns an empty array', () => {
-  //   const rawData = { kind: 'books#volumes', totalItems: 0 }
-  //   const bookData = BookDataTransformer.parseBookInfoList(rawData.items)
-  //   expect(bookData).toEqual([])
-  // })
 })
